Add XBN logo and info (#5651)
* Add XBT tokens information

Hello,

I'm proposing the new tokens to the crypto space. It is called Elastic Bitcoin or XBT.

XBT is a DeFi protocol with an elastic supply model. This means the total supply of XBT is constantly changing and achieved by utilizing the power of Ethereum's Smart Contract to adjust supply based on market condition.

* Delete info.json

* Delete logo.png

* Add XBT token information

Hello,

I'm proposing a new Defi protocol to the crypto space, which is called Elastic Bitcoin or XBT.

XBT is a DeFi protocol with an elastic supply model. This means the total supply of XBT is constantly changing and achieved by utilizing the power of Ethereum's Smart Contract to adjust supply based on market condition.

* add logo XBN and XBN tokens info

* add logo XBN and XBN tokens info

* add logo XBN and XBN tokens info

* add logo XBN and XBN tokens info

Co-authored-by: Alan The Achiever <alantheachiever@elasticbitcoin.org>