Add JailDoge Logo&Info. (#12343)
* Add JailDoge Logo&Info.

Jail Doge is a deflationary DeFi token that takes a 9% tax from each buy and sell. This tax acts to benefit the project as a whole by putting 3% in the liquidity pool for a stable price floor, 3% proportionally reflected back to all holders of a JAILDOGE token and 3% added to a marketing/dev wallet to make sure that the project is able to succeed in the long term. Another large purpose of the tax is to decentivise large token holders from buying and selling quickly and damaging the price. This almost always effects the small token holders and we want a fair coin for all.

* Change Info desc

* Change info desc