Add Sombra  (BEP20) (#13204)
* SMBR token Logo and Json

Sombra - https://sombra.app - is uploading our 256x256 logo and our .json info file to be added to trust wallet.

* SMBR token

Sombra - https://sombra.app - is uploading our 256x256 logo and our .json info file to be added to trust wallet.

* SMBR - Sombra logo and Json upload

We are uploading the Sombra (SMBR) token logo as well as the .json file for all information. Thank you.

* Sombra - SMBR logo and json info upload re-attempt

Uploading the Sombra Network logo and Json. We have fixed the non checksummed and duplicate ID errors in the json file. Thank you so much. 

Please let us know if the payment is something we need to do as we were unfamiliar with that. Our token previously was Cheecoin and we have now moved contracts and re-branded with the same community and team. Cheecoin was already on Trust. We are fine to pay if needed but wanted to check that we met the requirements as SMBR prior to sending the non refundable amount. 

Cheers,
Brendan and Sombra

* SMBR - Sombra Token json info and logo upload 

-We are re-uploading our .json file with fixes to position 10 (SMBR)
-We are re-uploading 256x256 logo with transparency (SMBR)

- Please remove https://github.com/Sombra-Network/assets/commit/2d3d91e5d5eac1b8bccbee63725acab5be089e2f#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6
- Please remove https://github.com/Sombra-Network/assets/commit/065ba4abd4a812fcced92c178ae370cdb73fc903#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6
- Please remove https://github.com/Sombra-Network/assets/commit/52aaf6089ff7791592265f6568021d47309d907c#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6

Committed by Brendan O

* Add files via upload

SMBR - Sombra Token json info and logo upload
-We are re-uploading our .json file with fixes to position 10 (SMBR)
-We are re-uploading 256x256 logo with transparency (SMBR)

- Please remove 2d3d91e#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6
- Please remove 065ba4a#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6
- Please remove 52aaf60#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6

https://github.com/Sombra-Network/assets/commit/279638edfeda5431ffc59bc27e4ae36036af7e48Committed by Brendan O

* Fixes for SMBR json and Logo - SOMBRA NETWORK

SMBR - Sombra Token json info and logo upload
-We are re-uploading our .json file with fixes to position 10 (SMBR)
-We are re-uploading 256x256 logo with transparency (SMBR)

- Please remove 2d3d91e#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6
- Please remove 065ba4a#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6
- Please remove 52aaf60#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6

* SMBR (Sombra Network) - json re-upload + logo

SMBR - Sombra Token json info and logo upload -We are re-uploading our .json file with fixes to position 10 (SMBR) -We are re-uploading 256x256 logo with transparency (SMBR). We uploaded a pull earlier but it seems to have errors so we will pay through this one.

	•	Please remove 2d3d91e#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6
	•	Please remove 065ba4a#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6
	•	Please remove 52aaf60#diff-41d4f38df2e41debc6b03a5036d4e91c606ff3c22c494c78b25684a0f8ed85e6
	•	Please remove 23ff3d6
	•	Please remove 279638e
	•	Please remove https://github.com/trustwallet/assets/pull/13183

* Delete info.json

* Delete logo.png

* Update info.json

* Delete logo.png

* Delete info.json

* Info for token Sombra

Co-authored-by: CryptoA <kulishnastia200@gmail.com>