Update TOKEN SDC  refer #17487 (#17743)
* Add token SDC 

Hi 
i' want request coin information for verification
If there is any method or extra cost to make it through urgently, please let me know.

Best Regard,

* Update Token SDC

Hi 
i' want request coin information for verification
If there is any method or extra cost to make it through urgently, please let me know.

Best Regard,

* Update Token SDC

Update Token SDC

* Update Token SDC

Update Token SDC Link

* Update Token SDC

Update Link

* Update Token SDC

Update SDC link

* Update SDC token

Update Link Sdc Info.json

* Update Token SDC

Update Link

* Update Token Sdc

Add Tag and Update Link

* Delete info.json

* Delete logo.png

* Update info.json

* Update TOKEN SDC  refer #17487

We have not reviewed it because it takes a long time and is very costly. Can we continue in the Trust because the cost has already been paid

* Update TOKEN SDC  refer #17487

We have not reviewed it because it takes a long time and is very costly. Can we continue in the Trust because the cost has already been paid

* Update TOKEN SDC  refer #17487

We have not reviewed it because it takes a long time and is very costly. Can we continue in the Trust because the cost has already been paid

* Update TOKEN SDC  refer #17487 

We have not reviewed it because it takes a long time and is very costly. Can we continue in the Trust because the cost has already been paid

* Delete info.json

* Delete logo.png

* Update TOKEN SDC refer #17487

We have already audited.

* Update info.json

Co-authored-by: Nazar <85495047+cryptomanz@users.noreply.github.com>