Follow up request for the request #21197 (#21316)
* Submitting the ENG token logo

Our project is a metaverse gaming platform and we deployed our gaming token on 2022/01/06. We have held our presale and ICO stages until now and we have reached 10,276 holders and 18,710 transactions.
You can check all the information through our bscscan official page.
This request is to submit our token logo to the trust wallet because most of our holders use this wonderful crypto wallet to hold their assets.

* Submit the ENG token logo

Our project is a metaverse gaming platform and we deployed our gaming token on 2022/01/06. We have held our presale and ICO stages until now and we have reached 10,276 holders and 18,710 transactions. You can check all the information through our bscscan official page. This request is to submit our token logo to the trust wallet because most of our holders use this wonderful crypto wallet to hold their assets.

* Delete info.json

* Delete logo.png

* Update info.json

* Information corrected for ENG token logo

We added the information that you requested to our info.json file.
payment status: paid

* Update info.json

* Update info.json

* Update info.json

* Update ENG token logo info:number of holders added

We have added the number of holders manually according to your request this time.

* Update info.json

* Update info.json

* Update info.json

* Update info.json

* Follow up request for previous request #21197

Hello again.
This is to follow up on the "Submit the ENG token logo #21197" request for the Endless Board Game project.
You requested us to provide an official SM audit and we did it with dessertswap.finance which you've suggested. 
We've provided all the necessary information in request #21197 and our payment status is "paid" and we've paid the fee.
The official link to our audit is as follows:
https://dessertswap.finance/audits/EndlessBoardGame-AUDIT-19175219.pdf

Thanks for your time and effort.
Regards

* Update info.json

* Update info.json

* Update info.json

* Update info.json

* Update info.json

* Update info.json

Co-authored-by: Nazar <85495047+cryptomanz@users.noreply.github.com>
Co-authored-by: CryptoA <82717856+Cryptocool1@users.noreply.github.com>